AE Submission Code/requirements.txt
      accelerate==0.22.0
      aiofiles==23.2.1
      aiohttp==3.8.5
      aiosignal==1.3.1
      altair==5.1.1
      anyio==3.7.1
      ##aspose-words
      async-timeout==4.0.3
      asyncio==3.4.3
      attrs==23.1.0
      beautifulsoup4==4.12.2
      bert-score==0.3.13
      binaryornot==0.4.4
      bs4==0.0.1
      cachetools==5.3.1
      certifi==2023.7.22
      chardet==5.2.0
      charset-normalizer==3.2.0
      click==8.1.7
      cmake==3.27.4.1
      colorama==0.4.6
      complexity==0.9.1
      contourpy==1.1.0
      cycler==0.11.0
      datasets==2.14.5
      dill==0.3.7
      distro==1.9.0
      exceptiongroup==1.1.3
      fastapi==0.103.1
      ffmpy==0.3.1
      filelock==3.12.3
      fonttools==4.42.1
      frozenlist==1.4.0
      fschat==0.2.26
      fsspec==2023.6.0
      google-ai-generativelanguage==0.4.0
      google-api-core==2.11.1
      google-auth==2.22.0
      google-generativeai==0.3.1
      googleapis-common-protos==1.60.0
      gradio==3.42.0
      gradio-client==0.5.0
      grpcio==1.58.0
      grpcio-status==1.58.0
      h11==0.14.0
      html2image==2.0.4.3
      htmlwebshot==0.1.2
      httpcore==0.17.3
      httpx==0.24.1
      huggingface-hub==0.16.4
      idna==3.4
      imgkit==1.2.3
      importlib-resources==6.0.1
      jinja2==3.1.2
      jiter==0.8.2
      jsonschema==4.19.0
      jsonschema-specifications==2023.7.1
      kiwisolver==1.4.5
      lit==16.0.6
      lxml==4.9.3
      markdown-it-py==3.0.0
      markdown2==2.4.10
      markupsafe==2.1.3
      matplotlib==3.7.2
      mdurl==0.1.2
      mpmath==1.3.0
      multidict==6.0.4
      multiprocess==0.70.15
      networkx==3.1
      nh3==0.2.14
      numpy==1.25.2
      openai==1.58.1
      opencv-python==4.8.1.78
      orjson==3.9.5
      packaging==23.1
      pandas==2.1.0
      peft==0.5.0
      pillow==10.0.0
      portalocker==2.7.0
      prompt-toolkit==3.0.39
      proto-plus==1.22.3
      protobuf==4.24.2
      psutil==5.9.5
      pyarrow==13.0.0
      pyasn1==0.5.0
      pyasn1-modules==0.3.0
      pydantic==1.10.12
      pydub==0.25.1
      pygments==2.16.1
      pyparsing==3.0.9
      python-dateutil==2.8.2
      python-multipart==0.0.6
      pytz==2023.3.post1
      pyyaml==6.0.1
      referencing==0.30.2
      regex==2023.8.8
      requests==2.31.0
      rich==13.5.2
      rouge==1.0.1
      rpds-py==0.10.2
      rsa==4.9
      sacrebleu==2.3.1
      safetensors==0.3.3
      scipy==1.11.2
      seaborn==0.13.0
      semantic-version==2.10.0
      sentencepiece==0.1.99
      shortuuid==1.0.11
      six==1.16.0
      sniffio==1.3.0
      soupsieve==2.5
      starlette==0.27.0
      svgwrite==1.4.3
      sympy==1.12
      tabulate==0.9.0
      tiktoken==0.4.0
      tokenizers==0.13.3
      toolz==0.12.0
      torch==2.0.1
      tqdm==4.66.1
      transformers==4.33.1
    
      typing-extensions==4.12.2
      tzdata==2023.3
      urllib3==1.26.16
      uvicorn==0.23.2
      wavedrom==2.0.3.post3
      wcwidth==0.2.6
      websocket-client==1.7.0
      websockets==11.0.3
      xxhash==3.3.0
      yarl==1.9.2
      zipp==3.16.2